+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes
 
+## Upcoming
+
+- Use correct error code "invalid_grant" when refresh token is not valid.
+
 ## v0.2.3 - 2025-04-18
 
 - Add HTTP endpoint to revoke all tokens for a user.